systemtests: don't create cleanup fixture from cleanup script per default
The default behaviour was to call the cleanup script as cleanup fixture
in every test. This unfortunately removes all logs and other information
that could be used to find the problem in case of a failure.

Now no cleanup fixture is created per default, so that in case of a
failure the logs and other information stays in the filesystem.

As the cleanup is called in the test-setup fixture before the setup
call, every test still gets a clean environment on startup so multiple
calls of the same test will not show any problems.